requirements.txt
streamlit
matplotlib
pandas_ta
mplfinance
smartapi-python
websocket
vectorbt
python-dateutil
requests
